【front】message型をmessage.d.tsの単一ファイルに統合
