+++ frontend/src/components/templates/media/chat/detail.tsx
@@ -2,8 +2,7 @@ import { useState, useRef, useEffect, useCallback, useMemo, FormEvent } from 're
 import { useRouter } from 'next/router'
 import { SubscribeIn } from 'types/internal/channel'
 import { ChatDetailOut } from 'types/internal/media/output'
-import { MessageCreateIn, MessageUpdateIn } from 'types/internal/message/input'
-import { ChatMessage, ChatReply } from 'types/internal/message/output'
+import { ChatMessage, ChatReply, MessageCreateIn, MessageUpdateIn } from 'types/internal/message'
 import { LikeMediaIn } from 'types/internal/user'
 import { postSubscribeChannel } from 'api/internal/channel'
 import { postMessage, getReplies, putMessage, deleteMessage } from 'api/internal/media/message'
