+++ frontend/src/types/internal/message.d.ts
@@ -1,5 +1,16 @@
 import { Author } from 'types/internal/user'
 
+export interface MessageCreateIn {
+  chatUlid: string
+  text: string
+  parentUlid?: string
+}
+
+export interface MessageUpdateIn {
+  chatUlid: string
+  text: string
+}
+
 export interface ChatMessage {
   ulid: string
   text: string
